move pts mutator interface and implementation into pkg, move quantity libaray into pkg also
