+++ pkg/mutators/podtemplatespec_test.go
@@ -4,8 +4,8 @@ import (
 	"os"
 	"testing"
 
-	"github.com/kanopy-platform/hedgetrimmer/internal/quantity"
 	"github.com/kanopy-platform/hedgetrimmer/pkg/limitrange"
+	"github.com/kanopy-platform/hedgetrimmer/pkg/quantity"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
